+++ super_test_engine/lib/super_test_engine/copy_app_templates/controllers/members_controller.rb
@@ -67,6 +67,14 @@ module Admin
           fields[:updated_at] = type.timestamp
         end
       end
+
+      def sortable_columns
+        ["id"] + super
+      end
+
+      def default_sort
+        { id: :asc }
+      end
     end
   end
 end